Desafio reactjs (#1)
* :art: refactor to desafio reactjs

* :alien: include Postman requests

+++ src/database/migrations/1588032513405-CreateTableTransaction.ts
@@ -29,7 +29,7 @@ export default class CreateTableTransaction1588032513405
           },
           {
             name: 'value',
-            type: 'Integer',
+            type: 'float',
           },
           {
             name: 'category_id',
